State Management Makes Life Easy
